requirements.txt
asgiref==3.5.2
Django==4.1.1
django-colorfield==0.7.2
djangorestframework==3.14.0
Markdown==3.4.1
Pillow==9.2.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
